+++ wait-while-starting.sh
@@ -1,6 +1,0 @@
-sleep 1
-while [ -f .starting ]
-do
-     echo "wait until build finished..."
-     sleep 1
-done