+++ packages/admin-color-picker/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [next]
+
+### Incompatible Changes
+
+-   Removed `clearButton` and `clearIcon` classes from color-picker
+    -   Using theme-augmentation the new common clear-button can now be styled with `CometAdminClearInputButton` instead of `VPAdminColorPicker`
+
+## [1.0.2] - 23. Feb 2021
+
+use fixed version of react-color
+
 ## [1.0.1] - 12. Jan 2021 - re-release under new name
 
 This package has been renamed to @comet/admin-color-picker